--- src/components/breadcrumbs/Breadcrumbs.module.css
@@ -1,18 +1,0 @@
-.container {
-  position: fixed;
-  display: flex;
-  left: 200px;
-  z-index: 1;
-  padding: 1rem 0 1rem 2rem;
-  width: 100%;
-  background-color: var(--grey-100);
-}
-
-.link {
-  text-decoration: none;
-  color: var(--grey-10);
-}
-
-.link:hover {
-  text-decoration: underline;
-}
