+++ src/components/navbar/NavBar.tsx
@@ -19,7 +19,7 @@ import { useRouter } from "next/router";
 import * as React from "react";
 import { MouseEventHandler } from "react";
 import $navbar from "./Navbar.module.css";
-import BreadcrumbsNav from "../breadcrumbs/BreadcrumbsNav";
+import BreadcrumbsNav from "../design_system/breadcrumbs/Breadcrumbs";
 
 interface NavItemProps {
   href?: string;
@@ -86,6 +86,7 @@ export default function NavBar() {
     );
   };
 
+  // Navigation Links
   const drawer = (
     <div className={$navbar.navbarDropdown}>
       <List>
